worker/requirements.txt
daemonize
Flask
Flask-PyMongo
pymongo
Pyphen
python3-pika
